correction type variables

+++ exercicesP1Unicap/src/higherNumber.java
@@ -2,9 +2,9 @@ import java.util.Scanner;
 public class higherNumber {
     public static void main(String[] args) throws Exception {
         Scanner sc = new Scanner(System.in);
-        int num1 = sc.nextInt();
-        int num2 = sc.nextInt();
-        int num3 = sc.nextInt();
+        double num1 = sc.nextInt();
+        double num2 = sc.nextInt();
+        double num3 = sc.nextInt();
         
         if(num1 >= num2 && num1 >= num3 && num2 >= num3){
             System.out.printf("%f >= %f >= %f", num1, num2, num3);
